+++ Lib/Platform/STM32G474xE/History.md
@@ -1,1 +1,6 @@
-Used for ECU25
+The following platform has been used for:
+- ECU25
+- ACU25
+- CCU25
+- GRFOCER (2025)
+- SAM (2025)